refactor(pathsafe): split DryRunLiveParity to drop cognitive complexity 18→<15 + clean Lane A naming
- 抽出 forceOverwriteParityCase + runForceOverwriteParityCase + setupRegularFile/Symlink/Directory
  helper，让 TestWritePlannedFiles_ForceOverwrite_DryRunLiveParity 从 18 跌到 ~3，
  满足 CLAUDE.md「函数认知复杂度 ≤ 15」上限（Sonar S3776 brain-overload Critical）。
- 合并 pathsafe_lane_a_test.go 全部 17 个测试到 pathsafe_test.go，并删除前者；
  "Lane A" 是 plan-041 (PR #526 已合并) 的 PM 临时车道命名，非 Go 测试惯例
  按特性/被测对象命名，已是已完工的死信息。
- pathsafe_internal_test.go godoc 同步去掉 "Lane A" / "A9" / "#544" 等 PM 编号字样。
- pathsafe.go ForceOverwrite 字段 godoc 去掉 "pre-A-API behavior" 的历史指代。

ref: gocognit/gocognit (cognitive complexity threshold)
Refs: Sonar S3776 (brain-overload Critical, 4-of-4 lift)
